MAPREDUCE-1508. Protect against NPE in TestMultipleLevelCaching. Contributed by Aaron Kimball
git-svn-id: https://svn.apache.org/repos/asf/hadoop/mapreduce/trunk@925554 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/mapred/org/apache/hadoop/mapred/TestMultipleLevelCaching.java
@@ -119,8 +119,11 @@ public class TestMultipleLevelCaching extends TestCase {
     		  testName, mr, fileSys, inDir, outputPath, 1, 1, 0, 0);
       mr.shutdown();
     } finally {
-      fileSys.delete(inDir, true);
-      fileSys.delete(outputPath, true);
+      if (null != fileSys) {
+        // inDir, outputPath only exist if fileSys is valid.
+        fileSys.delete(inDir, true);
+        fileSys.delete(outputPath, true);
+      }
       if (dfs != null) { 
         dfs.shutdown(); 
       }
